refactor: re-locate settings
